+++ wire/umsgblock.go
@@ -72,11 +72,11 @@ func BlockToAddLeaves(
 	remember []bool,
 	skiplist []uint32,
 	height int32,
-	outCount int) (leaves []accumulator.Leaf) {
+	outCount uint32) (leaves []accumulator.Leaf) {
 
 	// We're overallocating a little bit since all the unspendables
 	// won't be appended. It's ok though for the pre-allocation savings.
-	leaves = make([]accumulator.Leaf, 0, outCount-len(skiplist))
+	leaves = make([]accumulator.Leaf, 0, outCount-uint32(len(skiplist)))
 
 	var txonum uint32
 	for coinbaseif0, tx := range blk.Transactions() {
